requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
openai>=1.3.0
httpx>=0.25.0
pydantic>=2.5.0
weasyprint>=60.1
jinja2>=3.1.2
tqdm>=4.66.0
tenacity>=8.2.3
beautifulsoup4>=4.12.2
lxml>=4.9.3
serpapi>=0.1.5
crossrefapi>=1.6.0
rapidfuzz>=3.6.1
requests



